+++ src/QueryCat.Backend/Ast/Nodes/Select/SelectColumnsSublistAllNode.cs
@@ -5,8 +5,6 @@ internal sealed class SelectColumnsSublistAll : SelectColumnsSublistNode
     /// <inheritdoc />
     public override string Code => "column_all";
 
-    public static SelectColumnsSublistAll Instance { get; } = new();
-
     public SelectColumnsSublistAll()
     {
     }